refactor(connections): move credentials service to shared
