+++ packages/server/lib/controllers/mcp/connections/get.unit.test.ts
@@ -1,13 +1,13 @@
 import { afterEach, describe, expect, it, vi } from 'vitest';
 
+import { connectionCredentialsService, GetConnectionError } from '@nangohq/shared';
 import { Err, Ok } from '@nangohq/utils';
 
-import { connectionCredentials, GetConnectionError } from '../../../services/connectionCredentials.js';
 import { PublicMcpError } from '../utils.js';
 import { getConnectionsTool } from './get.js';
 
-import type { RetrievedConnection } from '../../../services/connectionCredentials.js';
 import type { ManagementMcpContext } from '../managementTool.js';
+import type { RetrievedConnection } from '@nangohq/shared';
 
 describe('getConnectionsTool', () => {
     afterEach(() => {
@@ -15,7 +15,7 @@ describe('getConnectionsTool', () => {
     });
 
     it('maps refresh arguments and omits credentials with the read scope', async () => {
-        const getSpy = vi.spyOn(connectionCredentials, 'get').mockResolvedValue(Ok(connectionFixture()));
+        const getSpy = vi.spyOn(connectionCredentialsService, 'get').mockResolvedValue(Ok(connectionFixture()));
 
         const result = await getConnectionsTool.handler(
             {
@@ -33,6 +33,8 @@ describe('getConnectionsTool', () => {
             environment: expect.objectContaining({ id: 42 }),
             connectionId: 'connection-id',
             integrationId: 'github',
+            onRefreshFailed: expect.any(Function),
+            onRefreshSuccess: expect.any(Function),
             returnRefreshToken: true,
             forceRefresh: true,
             refreshGithubAppJwtToken: true
@@ -45,7 +47,7 @@ describe('getConnectionsTool', () => {
     });
 
     it('returns credentials with the credential-reading scope', async () => {
-        vi.spyOn(connectionCredentials, 'get').mockResolvedValue(Ok(connectionFixture()));
+        vi.spyOn(connectionCredentialsService, 'get').mockResolvedValue(Ok(connectionFixture()));
 
         const result = await getConnectionsTool.handler(
             { connection_id: 'connection-id', integration_id: 'github' },
@@ -59,7 +61,7 @@ describe('getConnectionsTool', () => {
     });
 
     it('rejects invalid arguments before calling the connection service', async () => {
-        const getSpy = vi.spyOn(connectionCredentials, 'get');
+        const getSpy = vi.spyOn(connectionCredentialsService, 'get');
 
         const result = await getConnectionsTool.handler({ connection_id: 'connection-id' }, context(['environment:connections:read']));
 
@@ -76,7 +78,7 @@ describe('getConnectionsTool', () => {
         ['not_found', 'Connection does not exist'],
         ['invalid_credentials', 'Credentials are invalid']
     ] as const)('maps %s service errors to public MCP errors', async (code, message) => {
-        vi.spyOn(connectionCredentials, 'get').mockResolvedValue(Err(new GetConnectionError({ code, message })));
+        vi.spyOn(connectionCredentialsService, 'get').mockResolvedValue(Err(new GetConnectionError({ code, message })));
 
         const result = await getConnectionsTool.handler(
             { connection_id: 'connection-id', integration_id: 'github' },
